requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
supabase==2.9.1
neo4j==5.25.0
numpy==1.26.4
huggingface-hub==0.25.2
